apply suggested changes
