+++ src/tensor/mod.rs
@@ -132,7 +132,9 @@ pub use storage_traits::{AsArray, AsVec, CopySlice, TensorFromArray};
 pub use storage_traits::{DeviceStorage, HasErr};
 pub use storage_traits::{OnesTensor, SampleTensor, ZerosTensor};
 
-pub use tensor_impls::{PutTape, SplitTape, Tensor};
+#[cfg(feature = "cuda")]
+pub use tensor_impls::OnCuda;
+pub use tensor_impls::{OnCpu, OnDevice, PutTape, SplitTape, Tensor, ToDevice};
 pub use tensor_impls::{Tensor0D, Tensor1D, Tensor2D, Tensor3D, Tensor4D, Tensor5D, Tensor6D};
 
 #[cfg(test)]
